Don't add EDB to {apt,yum}_repository_list if it is defined
If the user explicitly defines the respective repository list ([] or
otherwise), we shouldn't add EDB to it, so we modify the default list
and use it only if one is not specified in config.yml.

Also tweak the EDB condition for Debian, which had drifted from the
analogous RH one, and avoid a warning about postgres_installation_src
not being defined.

Fixes #147