Don't install pglogical/pgespresso with BDRv1
Although BDRv1 passed EOL long ago, we must still support it in order to
test migrations to BDRv3. At the moment, it cannot be installed because
the pgespresso packages are missing and the pglogical packages cause a
conflict by depending on postgresql-9.4 (Debian), so we avoid them.

Reported by Gianni.