Add a |physical_replication_group filter
Given the name of a postgres instance, this filter returns a list of
names of all instances that are related to it by physical replication.

In this context, this means the primary and all of its streaming
replicas and their cascaded streaming replicas. Note that we set up
replicas to use streaming replication and do not support log-shipping
configurations, so although in theory "related to" encompasses both
forms of replication, here it implies streaming replicas only.

This filter is needed by the pgbackrest configuration.