Delay checking for EDB_REPO_CREDENTIALS_FILE
If we do the check in roles/platforms/common, we have to engage in some
convoluted logic to determine whether EDB will be in apt_repository_list
or yum_repository_list later, partly duplicating the logic elsewhere to
add it to those lists.

Moving it to roles/sys/pkg with the OS-specific repository setup tasks
makes the logic straightforward, at the expense of issuing the warning
(only slightly) later during the deploy.