Be more careful about quoting in shell tasks
In response to Lætitia's report of an error in the “Generate ssh
keypair” task caused by a space in the expansion of `inventory_file`
(the same thing would have happened in roles/secret as well).

This commit goes a little beyond fixing that problem—many of these are
unimportant because they're usually run on target instances with known
paths, but there's no harm in fixing them anyway. There might be a few
other cases that need fixing (e.g., build_path).