Add a new hosts_lines module to modify /etc/hosts
We switched from using lineinfile in a loop (agonisingly slow) to a new
linesinfile module that took a list of lines and operated on the file in
a single pass. However, its main shortcoming was that it couldn't remove
old entries for the same IP or hostname from /etc/hosts.

If you had an entry for `192.0.2.1 acid.example.com` and redeployed with
a custom `etc_hosts_lines` setting with a different address for the same
host, you would end up with two entries for the host. Bad.

Rather than adding some kind of regex-matching option to linesinfile, we
use a new module that takes a list of lines (as before), but knows how
to skip lines from /etc/hosts that refer to any IP or hostname that's
mentioned by one of the specified lines.

This quickly becomes a necessity if you're using `etc_hosts_lines` to
adapt to a complex network configuration, or if you need to change IP
addresses for some other reason.