Use `mapfile -t` to read lines into an array
I don't know how I ended up trying to do it with plain `read`; maybe I
meant to type `readarray` (an alias for mapfile).

The consequence of this error was that we only read the first line of
output, and therefore skipped various UNITS_TO_DISABLE.

Many thanks to Haroon for the analysis.