Move bdr_extensions from postgres_extensions to postgres_databases
So far, we added bdr_extensions to postgres_extensions and let the
extensions be created in template1 and inherited by bdr_database.
However, we want to avoid creating these extensions outside the
bdr_database, per advice from the BDR developers.

Now we use the ability to specify extensions in postgres_databases to
add the bdr_extensions to the default_bdr_database_entry; this way, it
will not be created in template1, only in bdr_database.

If we don't add bdr(_extensions) to postgres_extensions in roles/facts
any more, the code in postgres/config that decides whether to put bdr
into shared_preload_libraries now has no reason to do so; this won't
work, so we add a special case for bdr based on the instance role.

(This commit introduces a new filter named 'dictify' in order to convert
bdr_extensions to the [{name: …}, {name: …}, …] syntax required for the
postgres_databases[*].extensions entry.)