Remove unused BDR extension from databases other than bdr_database
We used to create the bdr extension in template1 and let it be inherited
by other databases, including bdr_database, but we create it directly in
bdr_database now (following advice from the BDR developers).

On existing BDR clusters, we drop the extension in every database (apart
from bdr_database) that does not have any bdr node(s) created.