Override ip_address with openvpn_ip after setting up openvpn
We always used to do this, but it got lost somewhere in the shuffle with
default_etc_hosts_lines and the new hosts_lines module, with the result
that we didn't end up putting the openvpn IPs into /etc/hosts at all,
which obviously meant the cluster didn't work as expected.

We now reconstruct default_etc_hosts_lines at the end of sys/openvpn,
and before sys/hosts, so that the addresses in /etc/hosts are correct.

Reported by Haroon.