Create the bdr extension with "cascade"
Vibhor reported that if bdr_database == "edb" on an EPAS installation,
the deploy would fail to create the bdr extension because pglogical was
missing, despite being in extra_postgres_extensions as expected.

The problem was that the edb database already existed (being the default
database on EPAS, analogous to the postgres database), so our creating
the pglogical extension in template1 had no effect on it. We do handle
the postgres database specially and create extensions in it alongside
template1, but this special case did not apply to edb.

In any case, using "CREATE EXTENSION bdr CASCADE" will create pglogical,
so we can just do that by default. (It would be nice to remove the need
to add pglogical to extra_postgres_extensions alongside this change, but
that's not possible because it is what controls putting pglogical into
shared_preload_libraries right now.)