Make postgresql_query return timing information
In addition to returning the "results" and "rowcounts" arrays, the
module now also returns a "runtimes" array with the time taken to
execute each query (in fractional seconds, with up to millisecond
precision).

If you execute only one query, it will also return "runtime", exactly as
it used to do with rowcounts/rowcount already.

Requested by Amruta.