Delay granting roles to users until after extensions are created
We moved bdr out of postgres_extensions and into postgres_databases, but
these grants were still being run immediately after processing only the
former, and before processing the per-database extensions declared in
the latter.

So if you had a user (like check_queue) with `granted_roles` depending
on the presence of the BDR extension, this task would fail.