Fix relative imports in architectures/lib/tpaexec
lib/tpaexec/platform/x.py should do `from . import Platform` to import
Platform from …/platform/__init__.py; likewise, architecture.py should
do `from .platform import Platform`, and so on. This allows pyright to
resolve all the imports.

However, we can't do this for architectures/bdr_always_on.py, because
that is a symlink to architectures/BDR-Always-ON/configure, and since
the location is different, the same relative import doesn't work. I'll
need to do a bit more research to decide how to fix that (and to bring
the other configure scripts under the purview of pyright, which won't
check files that aren't named *.py).

Apart from that one issue, we're now pyright-clean according to the
current (non-strict) pyrightconfig.json.