Allow setting postgres_wal_dir
If you set `postgres_wal_dir: /some/path` in your config, we will now
invoke `initdb --waldir=/some/path`, which will make $PGDATA/pg_wal a
symlink to the given directory.

The path must be absolute, and it must be outside postgres_data_dir. The
code does not currently check for these conditions, so initdb will fail
if they are not met.

(If you are already setting postgres_initdb_opts to add -X/--waldir to
set a custom location for pg_wal, you can leave things the way they are,
or remove your custom setting and set postgres_wal_dir instead. Don't do
both together.)

If you deploy a cluster with some postgres_wal_dir setting (or none, for
the default location), you cannot change it and redeploy. We detect the
current pgdata_wal_dir in library/minimal_postgres_setup and complain if
it is defined (i.e., it's from a running cluster) and doesn't match the
explicit setting of postgres_wal_dir. In other words, once pg_wal is
created, this code won't try to move it for you.