Prevent creating config file if it missing
The default behaviour of ini_file module is creating
the file if it is missing and we want to avoid this
for config files we do not install or control directly.
By setting create:no flag we can prevent the creation
of the files that do not exist.

In addition to this, the file might exist but the section
in the file does not. We do not cover this problem here.

The issue is reported by Jimmy.

Fixes #195.