Convert all configure scripts to wrappers for architectures/xxx.py
Instead of putting both the class definition and the executable wrapper
in the same file, we move all of the Architecture subclass definitions
to lib/tpaexec/architectures/ along with bdr.py.

The individual configure scripts just import the relevant class and call
configure() on an object of that class. Not really interesting, which is
as it should be. (I haven't tried to consolidate them into a single
lib/configure.py or something.)

This also means we can get rid of the symlinked bdr_always_on.py, and
all of the import resolution is uniform as far as pyright is concerned
(earlier, the import was relative to the configure script, and not for
the bdr_always_on.py file).

Also, the configure scripts earlier escaped formatting and pyright
checking, now the bulk of the code is under lib and subject to both.