Don't fail if rhnplugin.conf doesn't exist
This commit just patches a failure introduced by the fix for #195, which
added `create: no` to various ini_file invocations. (I checked, and the
others are all for files that we do expect to exist.)

I don't know (a) why rhnplugin.conf was missing on this RHEL8.2 system,
(b) whether it's OK to create it (it might be), and therefore (c) if
this commit is really the appropriate fix.

But until we can investigate in detail, this change avoids the failure.