Export the `match` test as `matched_by`
Ansible 2.9 removes support for exporting Jinja2 tests as filters, as
announced with the Ansible 2.4 release. This means one can no longer
write `s|match(p)`, but must instead write `s is match(p)`… which I
cannot bring myself to do.

Hence the `matched_by` export of the existing function. One could argue
about whether a string matches a regex or a regex matches a string, but
in practice, it's not possible to choose one or the other, it has to be
both to make sense.