Add implicit primary role for bdr instances
If a `bdr` instance is not explicitly marked as a `replica` in role, it
must be a `primary`, i.e., an instance on which we need to run initdb
(whereas we would run `repmgr standby clone` on a physical replica).

This makes it simpler to describe the meaningful combinations of roles
for a BDR instance:

- `role: [bdr]` is an ordinary BDR instance

- `role: [bdr, readonly]` is a BDR logical standby instance (which joins
  the node group with `pause_in_standby` set), eligible for promotion to
  a full participant

- `role: [bdr, subscriber-only]` is a BDR subscriber-only instance,
  which receives replicated changes but does not publish them

- `role: [bdr, replica]` is a physical replica of a BDR instance (it
  must have `upstream` set, as for any replica); not often used.

The first three combinations above are all implicitly `primary`
instances, whereas the last one is not.

This change does not affect config files that already set bdr+primary in
role. It just allows us to omit `primary` for brevity (and we do so for
newly-generated configurations).