Remove the cluster_name.{key,crt} installation mechanism
This was not a well-designed feature, and there's no reason to support
it. I'm reasonably sure nobody is actually using it (because it's just
not very sensible), but we error out if the files exist in cluster_dir,
so that nobody is caught by surprise.

Instances with the certificate already installed are left alone.