Guard against postgres_wal_dir == postgres_data_dir/pg_wal
The actions taken when postgres_wal_dir was defined all assumed that it
would be set to point to somewhere outside postgres_data_dir. That's not
unreasonable per se, but I needed to set it to /opt/postgres/data/pg_wal
to work around the problem with `repmgr standby clone` not handling the
setting correctly.

Although the problem with repmgr is now fixed, it seems prudent to guard
against this condition on the primary as well, and it doesn't take much
to do so. Now if you happen to explicitly set postgres_wal_dir to the
default postgres_data_dir/pg_wal, it'll be treated as if it's not set.