URL-encode EDB_REPO_CREDENTIALS_FILE before interpolating into baseurl
We expect EDB_REPO_CREDENTIALS_FILE to contain `username:password`, and
fail otherwise (platforms/common), but before we interpolate this into
edb.repo's baseurl, we must URL-encode the username and password.

Reported by Martín, who mistakenly used xxx@enterprisedb.com as his
username (should have been xxx), and got an error from curl via yum:

    Curl error (3): URL using bad/illegal format or missing URL for
    https://xxx@enterprisedb.com:XXXXXXXXXXX@yum.enterprisedb.com/edb/redhat/rhel-8-x86_64/repodata/repomd.xml
    [Port number ended with 'B']

After this change, yum reports this instead:

    Errors during downloading metadata for repository 'edb':
      - Status code: 401 for https://xxx%40enterprisedb.com:yyy@yum.enterprisedb.com/edb/redhat/rhel-8-x86_64/repodata/repomd.xml (IP: 54.165.250.135)
    Error: Failed to download metadata for repo 'edb': Cannot download repomd.xml: Cannot download repodata/repomd.xml: All mirrors were tried

Unfortunately, only the last line is quoted in the task failure, so this
doesn't help as much as I had hoped. But still, we shouldn't be adding a
syntactically-invalid URL to the Yum configuration to begin with.