Restructure docs directory (#210)
`mkdocs` isn't very happy with our current docs directory structure.
Especially true when making live changes during `mkdocs serve ...`
which results in below error:

"ERROR    -  Config value: 'docs_dir'. Error: The 'docs_dir' should not
be the parent directory of the config file. Use a child directory instead
so that the 'docs_dir' is a sibling of the config file."

Another reason to restructure is that tpaexec.pdf is currently generated
in `tpaexec/pdf` directory instead of `tpaexec/docs/pdf`(which is not ideal).
This is because we now generate pdf via mkdocs and it does not allow pdf
build directory parallel to docs_dir (*.md) so we had to move it out of `docs`
directory in a previous commit. New directory for docs structure is:

tpaexec/docs/
tpaexec/docs/tpa.yml (mkdocs project configuration)
tpaexec/docs/src/*.md (and other artifacts)
tpaexec/docs/pdf/

Signed-off-by: Muhammad Haroon <muhammad.haroon@enterprisedb.com>