Invoke semanage/restorecon once for multiple paths in postgres/initdb
We had ended up with four separate nearly-identical blocks of SELinux
context-setting commands for different custom Postgres paths (e.g.,
postgres_data_dir, postgres_wal_dir, …).

These commands are slightly tricky to call correctly, and expensive to
boot, so in the interest of clarity and efficiency, we coalesce these
tasks into a single one.

The trick is to build a list of directories whose context we need to
declare (and set), as well as a list of fcontext commands that we can
feed to `semanage -i -` via stdin. On the other hand, restorecon will
happily accept a list of directories on the command-line.

This commit also fixes a serious problem in #05b482f, which passed
`postgres_tablespaces` directly to semanage/restorecon, instead of
passing in the respective locations of the entries in the array.

(Originally by Richard, but with extensive changes by me.)

Closes #223