Add support for selectable --layout templates for an architecture
With BDR-Always-ON becoming available in various flavours, we introduce
a `--layout xxx` configure option that can be used to select a template
named layout/xxx.yml.j2 instead of main.yml.j2, to define the cluster
topology. This option is accepted only if the architecture provides
such layout templates.

This is a lightweight alternative to introducing a bunch of new named
architectures that inherit from BDR-Always-ON but do nothing more than
supply a different default main.yml.j2.

We also declare the default BDR-Always-ON flavour by symlinking
templates/main.yml.j2 to layouts/platinum.yml.j2 (although the current
layout does not correspond exactly to the proposed platinum layout).