Accept TPA_VENV=/path/to/venv to use an existing venv
The original intention was to accept an already-activated venv, but we
had hardcoded $TPA_DIR/tpa-venv in a couple of places, so this didn't
actually work (even though `tpaexec setup --venv /path/to/venv` did).

Given the status quo that running tpaexec with a non-standard venv did
not work, we require setting TPA_VENV explicitly in order to activate
the new behaviour, to avoid surprises in existing installations where a
different venv may have been active (but ignored by tpaexec).

Closes #254