Fixing user and template bug with rsyslog config drop-in files
The sys/rsyslog role wants to create a config which writes logs, as the
postgres user, coming in when it's in the log-server role. However, the
file wasn't being templated as it was just a copy. I've moved it to a template
but now rsyslog doesn't start because the postgres user doesn't exist. So, to
fix it I've added a include_role on a new postgres/user/tasks/accounts.yml
creation before the restart. Problem is that we can't be sure postgres will
be installed on that node or that we want users created that early when it is.
Assuming this is the least impactful option.

The second issue was with files being created as a the adm user on Debian
based Operating Systems. The default for all operating systems is now
root.
I was unable to find an appropriate user for Debian but it appears root is
better than creating the adm user as this may conflict with other Debian
system accounts created with low uid/gid numbers. The adm user on RHEL has
UID 3, this is taken by the sys user on Debian. The base-passwd package in
Debian enforces consistency for users with a UID of 0-99. If we were to add
adm it would need to be in the range 100-999.

Closes #256
Closes #257