Handle the differences in BDR4 packaging
The package name has changed for BDR4, and has 3 distinct variants
instead of two. Which means we need to have additional differentiation
for postgres family between community postgres and postgres extended
(former 2qPG). As part of this change, all packages now reside in single
repository for all postgres flavors as well.

BDR4 also does not have the pglogical component so don't install it.

This together means we have to add bdr_version to cluster_vars so that
it can be used during provision/deploy not just during configure step.

There could be larger refactoring done now that we have bdr_version
in cluster_vars, but that's left for the future.