Move selftest invocation to init.yml
We included roles/selftest from roles/platforms, and used `delegate_to`
and `run_once` to perform the checks once on localhost. This worked, but
had the weird and unpleasant side-effect of assigning localhost's facts
to all of the delegating target instance(s).

Here we add a new localhost-only play to init.yml to accomplish the same
thing in a cleaner way, without delegate_to/run_once.