Make all BDR instances subscribe to the witness-only repset
It turns out that it's not enough for bdr+witness nodes to subscribe to
the witness-only replication set; the other nodes must subscribe to it
too, because filtering which changes are sent to other nodes works only
if the nodes have at least one replication set in common. The earlier
setup resulted in witness nodes breaking when they couldn't apply DML
involving objects that they had never received the DDL to create.

The new configuration makes "witness-only" a potentially confusing name
("why is my non-witness node subscribed to this replication set?"), but
because we hope that future versions of BDR will handle this situation
better, we retain the current name for convenience.