Refactor BDR setup into 3 steps: init, join, post-join
This is so that common tasks that are needed for next step are executed
on all nodes before that next step so that we can have multiple versions
of BDR in the same cluster for rolling upgrade.

This is primarily needed because the join has to be executed in sequence
on one of the nodes and if we don't execute the info gathering tasks
before on all nodes regardless of the version the join will not work.

(Committed with minor structural reorganisation by AMS.)