Install new ca-certificates package early on RedHat (RHBA-2021:3649)
The Letsencrypt root certificate was cross-signed with the DST Root CA
X3 certificate, which expired at the end of September, as explained in
https://letsencrypt.org/docs/dst-root-ca-x3-expiration-september-2021/

The PGDG repository used a Letsencrypt certificate, and yum was unable
to verify its certificate if a recent enough ca-certificates (with the
preferred ISRG Root X1 certificate) was not installed. On an existing
instance where this was a problem, yum would fail to install packages
even from other repositories, so we install the new ca-certificates
package beforehand with disablerepo/enablerepo suitably tweaked.

Closes #234