Revert to original narrow disablerepo setting
We don't necessarily know which repository the updated ca-certificates
package will come from (it's distribution-specific: updates on centos:7,
baseos on centos:8 and rhel-8-baseos-rhui-rpms on RHEL8, and who knows
what on OEL).

So instead of trying to disable everything and enable only the right
repository, we should disable the only repository that we know has a
problem, and take the risk of some other custom repository also having a
Letsencrypt certificate. If that's the case, *this* task will fail, and
that's preferable to having a mysterious package installation failure
later on (from failing to install the latest ca-certificates here).

(This was the approach Matt's original PR took.)