Allow provision to gather tpaexec version info in a shallow clone.
Currently tpaexec tries to determine tpa version by looking for a
`VERSION` file; which exists in a release source tarball, or by
`git describe` in case `VERSION` file is missing. Second scenario
is typically true for a source clone. However, `git describe` does
not work for shallow clones and fails with an error like:

  fatal: No names found, cannot describe anything

Changed it to `git describe --always` so it works in case of a
shallow clone as well. Of course it would only output a short hash
when run on a shallow clone which is better than an error.

(Error reported by Gulcin)

Signed-off-by: Muhammad Haroon <muhammad.haroon@enterprisedb.com>