Harden checks to verify provisioned docker container is useful.
tpaexec would sometimes error out while provisioning docker containers
with a missing `networks` attribute error which looks like this:

 fatal: [localhost]: FAILED! =>
     {"msg": "The task includes an option with an undefined variable. The error was: 'dict object' has no attribute 'networks'\n\nThe error appears to be in
  '/opt/EDB/TPA/platforms/docker/docker_container.yml': line 61, column 3, but may\nbe elsewhere in the file depending on the exact syntax problem.\n\nThe offending line appears to be:\n\n\n- set_fact:\n  ^ here\n"}

It could happen for a few different reasons but the usually observed
scenarios is when docker engine fails to assign sensible network config
to the provisioned container. Typically a `docker inspect` output in such
a scenario would show that that running container has empty IPAddress
field as depicted in the following output:

   "Gateway": "172.17.0.1",
   "GlobalIPv6Address": "",
   "GlobalIPv6PrefixLen": 0,
   "IPAddress": "",

This commit attempts to add a couple of checks to verify whether or not
the provisioned container a) exists and b) carries sensible
networks settings before we attempt to access them.

Signed-off-by: Muhammad Haroon <muhammad.haroon@enterprisedb.com>