Revise max_worker_processes and set a reasonable floor value. (314)
On a small cluster, e.g. Training arch with one node base value for
max_worker_processes is too small, so the default of 16 is used.

{"ansible_facts": {"base_max_worker_processes": "2"}}
{"ansible_facts": {"max_worker_processes": "16"}}

On a larger cluster like BDR-Simple, the base value is calculated to be
something larger so kept.

{"ansible_facts": {"base_max_worker_processes": "25"}}
{"ansible_facts": {"max_worker_processes": "25"}, "changed": false}

Note: EPAS sets `dbms_aq.max_workers: 4` by default, which fails if
max_worker_processes is smaller. This change sidesteps the problem by
setting a minimum value of 16, which has a negligible impact on resource
usage. (If we ever want to make dbms_aq.max_workers larger, we will need
to explicitly increase max_worker_processes to compensate.)

Closes #314

(With minor editorial changes by AMS.)