Add new net property and subnet generation for architectures (32)
This is used where a configuration has an encompassing network defined
in addition to the subnets. Where a "--subnet" arg is defined we still
split it further into a number of subnets equal to locations. We also do
some math to make sure we can adequately contain the hosts within nets
defined or generated. An example is where Azure Vnets have a network
range defined and then subnet objects are created within that range.

Tests have been written to cover the functionality of the net.py module
so best examples of its use can be seen there.
The net module is a rewrite of the perl script architectures/lib/subnets
in python. The primary interface is through the Network class:

>>> net = Network(cidr='10.0.0.0/24', size=2)
>>> net.subnets()
Subnets(10.0.0.0/24): 10.0.0.0/26 10.0.0.64/26 10.0.0.128/26

The Subnets class does not randomise the subnets by default but its use
enables that functionality in architecture.py. A new CLI switch can be
used to turn that off.

Ref #32