Fix case where zero postgres instances exist (333)
A user might remove roles from the config which would make the number of
postgres instances evaluate to 0. This causes an error during deploy
which we want to avoid.

Noting that this doesn't prevent the installation of postgres packages
and config. This fix just addresses a bug which didn't exist before,
making it so that postgres isn't installled or configured at all would
need more discussion and development.

Closes #333