Ensure permissions for rsyslog managed postgres log is correct
The log file may have been created as owned by the user and group
'syslog' due to Ubuntu having the `$PrivDropToUser` and
`$PrivDropToGroup` directives. In order to retain the user and group
as 'postgres' we need rsyslog to start as the 'root' user. Therefore,
we first start by disabling that directive and when postgres is
configured we change the ownership on the files. If the file does not
yet exist or already has correct ownership this step should be skipped
and not register a changed item in the play recap.

Closes #304