Ensure correct package name for repmgr on RedHat from BDR Ent 2q (#340)
If we select BDR enterprise 2q package repos for RedHat we need to
specify a different package name for repmgr `postgresqlNN-repmgr`.
In order to do that the lookup map from defaults now uses the var
name postgres_family_extended which holds the flavour 2q when selected.
The variables needed in the config.yml needs to be the following for
this to work:

cluster_vars:
  postgresql_flavour: 2q
  tpa_2q_repositories:
    - products/2ndqpostgres/release
    - products/bdr_enterprise_3_7/release

(or a relevant version of BDR)

Closes #340