Remove the dependency of downloading etcd packages from Github
HARP v2 requires etcd v3.5.0 or above, which was not available in the
default package repositories for any distribution. As a workaround
TPAexec was downloading and installing etcd v3.5.0 from Github.

However, now `harp-etcd` and `harp-etcdctl` packages are in
products/harp/release package repositories so we no longer need to
download the packages from Github.

Fixes #359.