Don't install harp on non-proxy, non-primary-candidate instances
1. Invoke harp/pkg on harp-proxy instances and on postgres instances
   that are in first_bdr_primary_candidates (therefore excluding any
   witness/subscriber-only/replica/readonly instances).

2. Invoke harp/final on the same proxy/candidate set of instances.

3. Exclude subscriber-only instances from the list of DCS endpoints
   (because subscriber-only nodes don't participate in consensus, though
   witness/logical standby nodes do)

Note that we _do_ invoke the harp role on all instances, because we need
to invoke postgres/createuser on the non-primary-candidate BDR instances
so that group join doesn't fail when trying to restore privileges. So we
don't install harp, but we generate a useless configuration on instances
where we're not going to run harp.