Always run harp-{manager,proxy} as postgres_{user,group}
The harp packages hardcode postgres/postgres in the service unit files.
In #345, we hardcode enterprisedb/enterprisedb when running EPAS. Here
we switch to always using postgres_{user,group}, so that it works with
custom settings, and without any flavour-specific conditions.

Includes some cosmetic fixes, and also makes the override conditional on
the instance role, so as to avoid generating harp-xyz.d overrides on any
instances where we don't have that particular service installed.

Closes #374.