Switch cluster_addressing → ec2_instance_reachability in ec2.ini.j2
To use ec2 instances with private IP addresses (`assign_public_ip: no`),
we must tell ec2.py to return private IPs for each instance by setting
`vpc_destination_variable: private_ip_address`.

We used to do this if `cluster_addressing` was set to `private`, but
that seems to be a leftover in the template after the introduction of
`ec2_instance_reachability` long ago (in 2017). This change updates the
template so that the inventory plugin returns the private addresses as
expected.

Closes #376