Drop pglogical extension/schema where it's not usable
In databases where pglogical.node exists but pglogical_version() fails,
we need to drop the pglogical extension and schema, because pglogical is
gone (i.e., removed from shared_preload_libraries).

We exclude bdr_database, because we need to leave the extension alone
for the BDR upgrade scripts to remove it during the upgrade (which
should have happened already by the time this code runs).

Closes #381