Fix first_bdr_primary override for new node in new bdr_node_group
If you took an existing BDR cluster and added a new instance with a
different bdr_node_group setting, the deploy would fail with the
following error:

    FAILED! => {"msg": "The task includes an option with an undefined
    variable. The error was: No first item, sequence was empty.\n\nThe
    error appears to be in
    '/opt/2ndQuadrant/TPA/roles/facts/tasks/update-defaults.yml': line
    9, column 3, but may\nbe elsewhere in the file depending on the
    exact syntax problem.\n\nThe offending line appears to be:\n\n\n-
    name: Override first_bdr_primary if required\n ^ here\n"}

This override task looked for _any_ candidate BDR instances without
taking into account the bdr_node_group setting for those instances.

When a new node was added with a unique bdr_node_group, it would arrive
here with first_bdr_primary_candidates set correctly (to itself), but
_current_bdr_candidates would be wrong (all the other instances), and
their intersection would be empty, leading to the error above.

This change ensures that we only consider instances as candidates if
they have the same bdr_node_group setting as the node for which we are
overriding first_bdr_primary.

At the same time, the change also fixes a more subtle bug: we were
checking if the defined bdr_database was in cluster_facts.databases for
the candidate instances, but we were not using *that* instance's value
of bdr_database, which could well be different from the setting on the
evaluating instance. Since we don't currently support activating BDR in
multiple databases (though BDR supports it), we can safely remove this
check: having a non-empty `bdr_databases` is sufficient to show that the
candidate has completed bdr initialisation.

Closes #151