Make sure 'ping' command is available on EFM nodes
EFM uses 'ping.server.command' to run ping against 'ping.server.ip'
specified in $clustername.properties and the default command used
for 'ping.server.command' is specified as `/bin/ping -q -c3 -w5`.
This assumes that `ping` command is available on target nodes which
may or may not always be true. So we add 'iputils' and 'iputils-ping'
for RHEL and Debian/Ubuntu distributions respectively to
'efm_dependency_packages'. If `ping` is not available on target nodes
for some reason, you will see errors similar to the following in the
efm logs:

```
2022-02-15 14:14:23 Using ssl mode: disable
2022-02-15 14:14:23 Starting agent.
2022-02-15 14:14:25 Continuing startup in standby mode.
2022-02-15 14:14:25 Agent logs will be written to /var/log/efm-4.2/m1test.log
2022-02-15 14:14:25 The virtual.ip property is not set, starting EFM with VIP support disabled.
2022-02-15 14:14:25 Could not reach the address 8.8.8.8. Please check the value of your ping.server.ip property.
```

Signed-off-by: Muhammad Haroon <muhammad.haroon@enterprisedb.com>