Create a new harp-postgres.target ⊃ {postgres,harp_manager}.service
We need to stop and start postgres and harp_manager together, so we
define a new target that both services are `PartOf=`. This way, when
postgres/restart restarts the target, we stop harp-manager, restart
postgres, and start harp-manager; and harp-manager doesn't start
Postgres as soon as it notices it's been stopped.

Closes #388