Make harp_proxy settings conditional on harp_proxy_mode
We don't want to emit pgbouncer configuration when using the builtin
proxy, and vice versa. Also, we must install the pgbouncer package when
using `harp_proxy_mode: pgbouncer`, because the harp-proxy package does
not have a pgbouncer dependency any more (so that it can be used in the
now-default builtin mode without pgbouncer).

Also changes a couple of default configuration settings, including using
Unix sockets to connect to Postgres on the local instance.