Generate harp configuration only on instances where we install harp
While making the installation of harp conditional (e.g., to avoid
installing it on subscriber-only nodes), I decided it was too late at
night to also make the config generation conditional (since there were
some tasks in harp/config that needed to be run even on instances where
we didn't want to install harp).

This cleanup avoids generating the configuration if we didn't install
the packages.