Change default_harp_proxy_mode from pgbouncer to builtin
We used to deploy harp-proxy alongside a "captive" pgbouncer, but we now
want to migrate to using the builtin proxy instead. For now, pgbouncer
is still supported if you set `harp_proxy_mode: pgbouncer`.

WARNING: This change applies to existing clusters

Normally, we would require explicit configuration to opt-in to changes
like this, but in this case there should be no loss of functionality,
and circumstances (rapid development, limited user base with active
engagements, etc.) allow us to migrate with minimal disruption to
existing clusters.