Add archive-logs command to fetch logs from cluster instances
Created initially to aid with debugging of cluster integration tests
but useful whenever quick access to log files is required.

To create a log directory and archive logs from instances, run

```
tpaexec archive-logs <cluster-dir>
```

A `log` directory under <cluster-dir> will be created with a date/time
stamped directory under that. Then an ansible play will read the cluster
configuration to determine node names and then run tar to make an
archive of the `/var/log` directory on each instance. The tar archive is
in gzip format and will contain a leading directory with the instance
name and the same timestamp as the initial directory created on the
control node where `tpaexec archive-logs` was ran. The tar archives are
then copied back to the timestamped directory on the control node and
finally the archive is removed from the instance.

```
<cluster-dir>/logs/
`-- 20220218T135731
    |-- keratin-logs-20220218T135731.tar.gz
    |-- ketchup-logs-20220218T135731.tar.gz
    |-- klutz-logs-20220218T135731.tar.gz
    `-- useful-logs-20220218T135731.tar.gz
```

Archive contents example
```
keratin-logs
|-- anaconda
|   |-- anaconda.log
|   |-- dbus.log
|   |-- dnf.librepo.log
|   |-- hawkey.log
|   |-- journal.log
|   |-- ks-script-ipdkisn0.log
|   |-- ks-script-jr03uzns.log
|   |-- ks-script-mh2iidvh.log
|   |-- lvm.log
|   |-- packaging.log
|   |-- program.log
|   |-- storage.log
|   |-- syslog
|   `-- X.log
|-- btmp
|-- dnf.librepo.log
|-- dnf.log
|-- dnf.rpm.log
|-- hawkey.log
|-- lastlog
|-- private
|-- tpaexec.log
`-- wtmp
```