Relocate python module tpaexec to lib (#393)
This is so that we can standardise on `lib/` being the location for
libraries. It will allow for other library based tooling to be used in a
uniform fashion, such as packaging/lint/test/etc.
Other scripts in architectures/lib are "binary" implementations and
should be considered separately for migration to a library layout
under lib if applicable.

We add $TPA_DIR/lib to PYTHONPATH, so that it is set correctly for
anything invoked by either ansible/ansible or _with_ansible_env (which
is meant to invoke things that aren't ansible, e.g., a configure script,
in the same way that ansible would be invoked).

Closes #393