Create S3 cluster bucket according to users' choice
This reverts and modifies commit #33ed84e9e185bee21f889dcaf2c11fba267db228
(Switch to using a single long-lived S3 bucket"), and provides the user
with an option to either create a bucket during provision or use an
existing one using the config option `cluster_bucket` or via a new
`--cluster-bucket` configure argument.

Previously, TPAexec was using the `2ndquadrant-tpa` bucket to store
SSH keys. This had the right permissions under 2ndQuadrant
AWS account. Now that we are provisioning instances under the EDB AWS
organisation where each user has their own sandboxed account.

Creating a single bucket that is accessible to every user in the Org
was decided against on the basis that each account would have a separate
global namespace for bucket naming unlike before with the 2Q account
and to avoid TPA needing to maintain Org level IAM policies for the
bucket.

There are some concerns over provisioning an S3 bucket on every run
for each cluster. Bucket synchronisation across regions is not
instantaneous leading to spurious errors when attempting various
operations ranging from setting lifecycle options or policies on buckets,
to the availability of objects. More detailed info can be found in the
original reverted commit above.

The following rules exist for the `cluster_bucket` config option:

* If the option is not set or if it is set to "auto" we create a bucket
  with a name based on the users' ID. The name is filtered accordingly;
  it cannot contain special characters (except hyphens) or upper-case
  letters, and needs to be less than 64 characters in length.

* Unless the option is set to "auto", we ask the user via a prompt to
  confirm creation of the bucket.

* If the option is set to anything else we check that a bucket of that
  name exists and if it does not we exit with an error.

We do not remove the bucket automatically during deprovision.
Instructions are provided on how to create and remove the bucket.

If we create the bucket we set expiry on all objects to 1 day.
To avoid any confusion with lifecycle management used in the aws_s3 task,
the unused `expiry` and `register` options have been removed.
The set-provisioning-var key name has also been renamed to avoid similar
confusion.

With extensive contributions and testing from Matt.

Closes #416